add freezer proxy to media

+++ services/services/freezer/ipc/freezer_service_proxy.cpp
@@ -47,7 +47,7 @@ int32_t FreezerServiceProxy::ProxyApp(const std::unordered_set<int32_t>& pidSet,
         data.WriteInt32(pid);
     }
     data.WriteBool(isFreeze);
-    int error = Remote()->SendRequest(FreezerServiceMsg::PROXY_APP, data, reply, option);
+    int error = Remote()->SendRequest(static_cast<int32_t>(FreezerServiceMsg::PROXY_APP), data, reply, option);
     if (error != MSERR_OK) {
         MEDIA_LOGE("ProxyApp failed, error: %{public}d", error);
         return error;
@@ -66,7 +66,7 @@ int32_t FreezerServiceProxy::ResetAll()
         return MSERR_UNKNOWN;
     }
 
-    int error = Remote()->SendRequest(FreezerServiceMsg::RESET_ALL, data, reply, option);
+    int error = Remote()->SendRequest(static_cast<int32_t>(FreezerServiceMsg::RESET_ALL), data, reply, option);
     if (error != MSERR_OK) {
         MEDIA_LOGE("ResetAll obj failed, error: %{public}d", error);
         return error;
