+++ services/services/freezer/ipc/freezer_service_stub.cpp
@@ -51,8 +51,8 @@ FreezerServiceStub::~FreezerServiceStub()
 
 int32_t FreezerServiceStub::Init()
 {
-    freezerFuncs_[FreezerServiceMsg::PROXY_APP] = &FreezerServiceStub::HandleProxyApp;
-    freezerFuncs_[FreezerServiceMsg::RESET_ALL] = &FreezerServiceStub::HandleResetAll;
+    freezerFuncs_[static_cast<int32_t>(FreezerServiceMsg::PROXY_APP)] = &FreezerServiceStub::HandleProxyApp;
+    freezerFuncs_[static_cast<int32_t>(FreezerServiceMsg::RESET_ALL)] = &FreezerServiceStub::HandleResetAll;
     return MSERR_OK;
 }
